+++ apps/studio/src/ui/global/widgets/components/ChallengesTable/ChallengesTableView.tsx
@@ -130,7 +130,7 @@ export const ChallengesTableView = ({
               </TableCell>
               <TableCell>{challenge.downvotesCount ?? 0}</TableCell>
               <TableCell>{challenge.upvotesCount ?? 0}</TableCell>
-              <TableCell>{challenge.completionsCount ?? 0}</TableCell>
+              <TableCell>{challenge.completionCount ?? 0}</TableCell>
               <TableCell>
                 {challenge.slug ? (
                   <Button variant='ghost' size='icon' asChild>