+++ packages/core/src/space/domain/entities/tests/fakers/PlanetsFaker.ts
@@ -16,7 +16,8 @@ export class PlanetsFaker {
       icon: `${faker.image.avatar()}.jpg`,
       image: `${faker.image.avatar()}.jpg`,
       stars: StarsFaker.fakeManyDto(faker.number.int({ min: 3, max: 10 })),
-      completionsCount: faker.number.int({ min: 0, max: 100 }),
+      completionCount: faker.number.int({ min: 0, max: 100 }),
+      userCount: faker.number.int({ min: 0, max: 100 }),
       isAvailable: faker.datatype.boolean(),
       ...baseDto,
     }